Split shell internal POSIX tests
